Tightened up current status.

+++ pokertournament-api/src/main/java/com/pilon/pokertournament/tournamentState/TournamentCurrentState.java
@@ -40,7 +40,7 @@ public class TournamentCurrentState {
 
     @Column
     @NotNull
-    private Integer durationRemainingSeconds;
+    private Long durationRemainingSeconds;
 
     @Column
     @NotNull
